Revisão final

+++ src/Estrela.java
@@ -8,7 +8,7 @@ public class Estrela extends CorpoCeleste implements Gravitacional {
     private SistemaSolar sistemaSolar;
     private List<Planeta> planetas;
 
-    public Estrela(String nome, double massa, double raio, int idade, String tipoDeEstrela, double temperatura, SistemaSolar sistemaSolar) {
+    public Estrela(String nome, double massa, double raio, double idade, String tipoDeEstrela, double temperatura, SistemaSolar sistemaSolar) {
         super(nome, massa, idade);
         this.raio = raio;
         this.tipoDeEstrela = tipoDeEstrela;
